% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rust_pwd}
\alias{rust_pwd}
\title{Resolve Full Path}
\usage{
rust_pwd(path)
}
\arguments{
\item{path}{A character string representing a file path to resolve}
}
\value{
A character string with the canonical absolute path
}
\description{
This function resolves a relative path to its canonical absolute representation.
}
\keyword{internal}
