% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{string_embedding}
\alias{string_embedding}
\title{only Bert models a currently supported and the following files are expected:
\itemize{
\item config.json
\item tokenizer.json
\item model.safetensors
}}
\usage{
string_embedding(x, path)
}
\description{
only Bert models a currently supported and the following files are expected:
\itemize{
\item config.json
\item tokenizer.json
\item model.safetensors
}
}
